Correctly handle the -- operator in spec strings
Placing a -- in a spec string is equivalent to the user providing the same operator in the program call arguments in the same position.

To implement this, the normalization phase (splitting folded options, auto-inserting of true for boolean options, etc.) had to be dropped altogether
because it wasn't aware of the state of the rejectOptions flag and thus it can reject valid calls.

Fixes #7