Rework the orderless options handling logic
Previously, mow.cli used to handle the orderless nature of options in
a, shall we say "not a very smart way" ?

The sequence parser, if two consecutive elements A and B only contain options,
a new FSM accepting "A then B" and "B then A" was constructed.

For 3 options-only elements, this would amount to 6 possible
combinations, 24 for 4 elemnts, etc.

The time and memory compexity of this rapidly grows on a factorial of
the number of options.

The proposed fix here attacks the problem the other way around:
instead of tweaking the FSM to accept the options in any order, the FSM
is kept as is, but matching an option would look over all the
arguments, and not just the first.

Given a spec string:

`-f -g`

The generated FSM would look like this:

`S0--[-f]-->S1--[-g]-->(S2)`

When applied to a user input of `-g -f`, the first transition (`S1-->S2`, which only succeeds if the `-f` option is encountered)
would check the args one by one until it finds a match, or fail
otherwise.

Fixes #28