Fix infinite loop case for envvar backed opts
optsmatcher should not fall back an environment variable as a
requirement for an option during a transition; doing so may cause
optsmatcher.match to never exit as the generated optmatcher will always
return true, even when it doesn't consume any arguments.

Generate a new optOrEnvMatcher for use in cases where falling back to an
environment variable satisfies the requirement condition.