Fix infinite loop case for envvar backed opts
This is a regression introduced by the feature "Allow required options to be satisfied by an nev var".

The bug lies in `optsMatcher`, which iterates over its options, and keeps looping until no more options can match or there is no more input to feed them.

If one of the options can be satsified from an en var, it doesn't consume any input, and returns true, causing an inifinite loop.

The proposed fix is to exclude an option (in `optsMatcher`) whenever  it has been satisfied by an env var.