Support rerouting output and exit code handlers.
While a CLI application most commonly does want to call os.Exit
to handle exit codes, and print errors or usage information to the
stderr file descriptor... it's also useful to be able to control these.

By allowing these to be user-controlled, it becomes possible for users
to write tests of their CLI behavior, _without_ needing to invoke
their entire program as a subprocess (which would carry considerable
complexity burdens; invoking go-install in the middle of go-test...
well, let's say it "does not spark joy").

Previously, within this library, the needs for testing were satisfied
by using package-scope variables, which aside from being global,
were also unexported.  This is no longer necessary.  The internal
tests now use the same mechanisms that will be available to users.
(As a bonus, the tests could now be parallelized without issue,
though I haven't taken that step in this commit.)