Publish `Stderr` to allow testing help
It is useful to test the help output for the non-trivial CLIs and in the
environment where an example outputs are desired.

Before this, testing for help required resorting to `os/exec`.

After this, the consumer test can swap the `cli.Stderr` for testing
purposes while not affecting any other details of the `mow.cli`.

The `stdOut` is not published as it is only used internally for
testing and has no impact on consumers as `mow.cli` never prints to
STDOUT.

NOTE: [alternative implementation] exists but has a larger "blast radius".

[alternative implementation]: https://github.com/jawher/mow.cli/pull/128